Removing mathbf from titles
Turns out using commands defined with mathbf in
entry titles cause the error:
Tex capacity exceeded

I suspect this has something to do with the fact
that chapter titles are defined using bfseries and
boldmath, and using mathbf may cause an infinite
loop.

I know this solution is less than ideal, but I
could not find a way to get mathbf to work in the
titles, and most of them were just an acronym.